requirements.txt
pillow
python-docx
